refactor(inventory): extract formatLiferayInventoryPage to format module
- Move formatLiferayInventoryPage (226 LOC) from liferay-inventory-page.ts
  to new liferay-inventory-page-format.ts
- Extract appendConfigurationTabsLines and appendFragmentEntryLines as
  private helpers (previously inline code)
- Remove eslint-disable max-lines comment (no longer needed)
- Add re-export of formatLiferayInventoryPage from main module
